+++ src/calc-serialized-size.h
@@ -1,0 +1,2 @@
+
+int calc_serialized_size(SEXP robj);